+++ src/hooks/profile_page.cpp
@@ -1,6 +1,9 @@
 #include "profile_page.hpp"
 
+#ifndef GLOBED_DISABLE_EXTRA_HOOKS
+
 #include <managers/admin.hpp>
+#include <net/manager.hpp>
 
 using namespace geode::prelude;
 
@@ -22,12 +25,22 @@ void GlobedProfilePage::loadPageFromUserInfo(GJUserScore* score) {
 	};
 
 	Build<CCSprite>::createSpriteName("GJ_reportBtn_001.png")
-	.scale(.7f)
-	.intoMenuItem([player](auto) {
-		AdminManager::get().openUserPopup(player.makeRoomPreview(0));
-	})
-	.parent(leftMenu)
-	.id("admin-button"_spr);
+		.scale(.7f)
+		.intoMenuItem([player](auto btn) {
+			auto& nm = NetworkManager::get();
+			if (!nm.established()) {
+				auto parent = btn->getParent();
+				btn->removeFromParent();
+				parent->updateLayout();
+				return;
+			}
+
+			AdminManager::get().openUserPopup(player.makeRoomPreview(0));
+		})
+		.parent(leftMenu)
+		.id("admin-button"_spr);
 
 	leftMenu->updateLayout();
 }
+
+#endif // GLOBED_DISABLE_EXTRA_HOOKS